requirements.txt
flask
python-dotenv==0.17.1
python-telegram-bot==13.4.1
requests
bitcoin
mnemonic

